requirements.txt
colorama==0.4.6
openai==1.82.1
python-dotenv==1.0.1 